Create ProductInfo component
