+++ src/scss/App.scss
@@ -98,6 +98,10 @@ body {
   line-height: 25px;
   letter-spacing: 2px;
   text-transform: uppercase;
+
+  &--orange{
+    color: $accent;
+  }
 }
 
 .button_orange {
